UUID: Stop caching clock_seq_and_node
Previously, clock_seq_and_node was generated once at startup and was used for
generation of TimeUUIDs. This is problematic because as a result creating
TimeUUID at the same time on the same node always creates exactly the same
value. It's bad because the whole purpose of using TimeUUID is to create
unique identifier for a timestamp. Not a duplicate. Otherwise a timestamp
alone could be used as well.

This patch drops the cache of clock_seq_and_node and calls
make_clock_seq_and_node() every time TimeUUID is generated. This creates new
unique clock_seq_and_node for each generation and reduces the chance for
collision.

Fixes #6208

Tests: unit(dev)

Signed-off-by: Piotr Jastrzebski <piotr@scylladb.com>